First step to restructure the code (libraries, file organization, etc.). New library : btk::Base. This library contains only the objects needed for the internal behaviour of the project.

+++ Code/Base/btkLogger.cpp
@@ -235,7 +235,7 @@ namespace btk
       return;
     if (this->mp_Pimpl->Output == nullptr)
       this->mp_Pimpl->Output = new Console;
-    size_t n = strlen(msg)*2;
+    int n = strlen(msg)*2;
     char* str = new char[n];
     while (1)
     {
